+++ CHANGELOG.md
@@ -17,9 +17,10 @@
 ### Additions and Improvements
 - Add blockTimestamp to receipt logs for `eth_getBlockReceipts` and `eth_getTransactionReceipt` results [#9294](https://github.com/hyperledger/besu/pull/9294)
 - Upgrade to execution-spec-tests v5.3.0 [#9301](https://github.com/hyperledger/besu/pull/9301)
-- Update netty [#9330](https://github.com/hyperledger/besu/pull/9330)
+- Update to netty 4.2.7.Final [#9330](https://github.com/hyperledger/besu/pull/9330)
 - Ability to enable/disable stack, storage and returnData tracing data in debug_traceStandardBlockToFile and debug_traceStandardBadBlockToFile endpoints [#9183](https://github.com/hyperledger/besu/pull/9183)
 - Increase mainnet gas limit to 60M [#9339](https://github.com/hyperledger/besu/pull/9339)
+- Update to tuweni 2.7.2 [#9338](https://github.com/hyperledger/besu/pull/9338)
 - Add `opcodes` optional parameter to RPC methods: `debug_standardTraceBlockToFile`, `debug_standardTraceBadBlockToFile`, `debug_traceBlockByNumber`, `debug_traceBlockByHash`, `debug_traceTransaction`, `debug_traceBlock`, `debug_traceCall` for tracing specified opcodes [#9335](https://github.com/hyperledger/besu/pull/9335)
 
 ### Bug fixes